requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.34
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
alembic==1.13.2
python-dotenv==1.0.1
loguru==0.7.2
google-genai
sounddevice>=0.4.6
soundfile>=0.12.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
dlib==19.24.6
docopt==0.6.2
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg==1.4
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
huggingface-hub==0.29.2
HyperPyYAML==1.2.2
idna==3.10
importlib_resources==6.5.2
ipython==8.18.1
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
julius==0.2.7
jupyterlab_widgets==3.0.13
keyboard==0.13.5
kiwisolver==1.4.7
lightning==2.5.0.post0
lightning-utilities==0.14.0
llvmlite==0.43.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
optuna==4.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
primePy==1.3
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==6.30.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==21.0.0
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
semver==3.0.4
sentencepiece==0.2.0
setuptools-rust==1.10.2
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.2
SQLAlchemy==2.0.38
srt==3.5.3
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizer==3.4.5
tokenizers==0.21.0
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.6.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vosk==0.3.45
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
